guard/requirements.txt
# Runtime dependencies
# 설치: python3 -m pip install -r requirements.txt

discord.py>=2.3.2
Pillow>=10.0.0
ImageHash>=4.3.1
numpy>=1.24.0
zxing-cpp>=2.2.0
